app/requirements.txt
gradio==3.41.2
google-cloud-speech>=2.24.0
